chore(consensus): move canister http tests into a separate lib
This change is a prerequisite step that will allow reuse of the tests
in a context that tests for cycles being refunded based on the spec.
