+++ rs/tests/networking/BUILD.bazel
@@ -73,6 +73,29 @@ system_test_nns(
     ],
 )
 
+rust_library(
+    name = "canister_http_correctness_tests_lib",
+    testonly = True,
+    srcs = [
+        "canister_http_correctness_tests_lib.rs",
+    ],
+    crate_name = "canister-http-correctness-tests-lib",
+    target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
+    deps = CANISTER_HTTP_BASE_DEPS + [
+        "//rs/registry/subnet_type",
+        "//rs/rust_canisters/canister_test",
+        "//rs/test_utilities",
+        "//rs/test_utilities/types",
+        "//rs/types/base_types",
+        "//rs/types/types",
+        "@crate_index//:assert_matches",
+        "@crate_index//:ic-agent",
+        "@crate_index//:serde",
+        "@crate_index//:serde_json",
+        "@crate_index//:tokio",
+    ],
+)
+
 system_test_nns(
     name = "canister_http_correctness_test",
     env = {
@@ -88,6 +111,7 @@ system_test_nns(
         "//rs/rust_canisters/canister_test",
         "//rs/test_utilities",
         "//rs/test_utilities/types",
+        "//rs/tests/networking:canister_http_correctness_tests_lib",
         "//rs/types/base_types",
         "//rs/types/types",
         "@crate_index//:assert_matches",
